+++ release/MetricsWeighted/R/AUC.R
@@ -1,6 +1,6 @@
 #' Area under the ROC
 #'
-#' Calculates weighted AUC, i.e. the area under the receiver operating curve, based on a deterministic version of \code{glmnet::auc}. The larger the AUC, the better.
+#' Function copied from glmnet package (modified to ensure deterministic results). Calculates weighted AUC, i.e. the area under the receiver operating curve. The larger, the better.
 #'
 #' The unweighted version can be different from the weighted one with unit weights due to ties in \code{predicted}.
 #'