+++ release/MetricsWeighted/R/mae.R
@@ -1,4 +1,4 @@
-#' Mean absolute error
+#' Mean Absolute Error
 #'
 #' Calculates weighted mean absolute error of predicted values. The smaller the value, the better.
 #'