+++ release/MetricsWeighted/R/weighted_mean.R
@@ -1,4 +1,4 @@
-#' Weighted mean
+#' Weighted Mean
 #'
 #' Returns weighted mean of a numeric vector. In contrast to \code{stats::weighted.mean}, \code{w} does not need to be specified.
 #'
@@ -12,7 +12,8 @@
 #' weighted_mean(1:10)
 #' weighted_mean(1:10, w = NULL)
 #' weighted_mean(1:10, w = 1:10)
-#' # weighted_mean(1, 0)
+#' # weighted_mean(1, 0) # Raises error
+#' @seealso \code{\link{weighted_quantile}}.
 weighted_mean <- function(x, w = NULL, ...) {
   if (is.null(w)) {
     return(mean(x, ...))